fix(indexer): SMI-6015 round-2 review fix + 500-line split
Round-2 GPT-5.6-Sol review of the previous commit's fixes found one real
issue: the dedicated 401 retry's attempts++ was shared with the outer
while loop's own retry-budget counter, so a 401-then-recovered-to-a-
retryable-status sequence (e.g. 401 -> 403 -> 5xx) silently lost one of
its outer retry attempts -- contradicting resolveOne's own documented
'dedicated, outer-loop-independent' contract. Split into two counters:
attempts (pure outer-loop budget, untouched by the 401 branch, exactly as
before the original fix) and totalRequests (the real HTTP request count,
what ResolutionOutcome.attempts actually reports). New regression test
walks the reviewer's exact 401->403->5xx scenario and asserts all 4
requests happen.

The attempts-counter fix (515 lines) pushed
smi5879-census.branches.helpers.ts over CLAUDE.md's 500-line-per-file
budget. Split the batched-write / SQL-builder machinery (no logical
boundary change, same code) into a new sibling
smi5879-census.branches.writes.ts, with its own test file split the same
way. All three files now well under budget (408/123/318 lines).

1 new test. Full indexer suite (1393 tests) green; lint, format,
typecheck, audit:standards all clean.

Refs SMI-6015
